requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.0
alembic==1.6.5
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
attrs==21.4.0
autopage==0.4.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
# cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1636046063618/work
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
cliff==3.10.0
cmaes==0.8.2
cmd2==2.3.3
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configparser==5.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
docker-pycreds==0.4.0
einops==0.4.0
entrypoints==0.3
filelock==3.4.2
flake8==4.0.1
fonttools==4.28.5
frozenlist==1.2.0
fsspec==2022.1.0
# future @ file:///home/conda/feedstock_root/build_artifacts/future_1635819504071/work
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.43.0
hydra-colorlog==1.1.0
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.2
identify==2.4.3
idna==3.3
imageio==2.13.5
imgaug==0.4.0
importlib-metadata==4.10.0
importlib-resources==5.2.3
iniconfig==1.1.1
ipykernel==6.6.1
ipython==7.31.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-server==1.13.1
jupyterlab==3.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.3
multidict==5.2.0
mypy-extensions==0.4.3
nbclassic==0.3.4
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nbstripout==0.5.0
nest-asyncio==1.5.4
networkx==2.6.3
nodeenv==1.6.0
notebook==6.4.6
numpy==1.22.3
oauthlib==3.1.1
# olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.1.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
optuna==2.10.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
prettytable==3.0.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.3
psutil==5.9.0
ptyprocess==0.7.0
pudb==2022.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
# pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyDeprecate==0.3.1
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.6
pyperclip==1.8.2
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
pytorch-lightning==1.6.4
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qudida==0.0.4
# ray==1.12.0
# ray-lightning==0.2.0
requests==2.27.1
requests-oauthlib==1.3.0
rich
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.5.2
setproctitle==1.2.2
sh==1.14.2
Shapely==1.8.0
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.29
stevedore==3.5.0
subprocess32==3.5.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
timm @ git+https://github.com/rwightman/pytorch-image-models.git@7c67d6aca992f039eece0af5f7c29a43d48c00e4
toml==0.10.2
tomli==1.2.3
torch
torchmetrics
# -e git+https://github.com/OZOOOOOH/imbalanced-dataset-sampler.git@9c4251e3c078d4db7f4120f447101f878a5215b5#egg=torchsampler
torchvision
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
# typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1638334978229/work
urllib3==1.26.8
urwid==2.1.2
urwid-readline==0.13
virtualenv==20.13.0
wandb
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
